feat：修改统计，分离emoji动作
